Extracting unit to its own component.
